requirements.txt
opencv-python==3.4.3.18
torch==1.7.1
torchsummary==1.5.1
torchvision==0.8.2
numpy==1.18.1
pandas==0.23.4
scikit-image==0.16.2
scikit-learn==0.23.2
jsonpickle==1.2
Pillow==7.1.1
matplotlib==3.0.2
